+++ src/Opaleye/PGTypes.hs
@@ -1,4 +1,8 @@
-{-# LANGUAGE EmptyDataDecls #-}
+{-# LANGUAGE EmptyDataDecls, CPP #-}
+
+#if defined(__GLASGOW_HASKELL__) && !defined(__HADDOCK__)
+#include "MachDeps.h"
+#endif
 
 module Opaleye.PGTypes (module Opaleye.PGTypes) where
 
@@ -17,7 +21,7 @@ import qualified Data.ByteString.Lazy as LByteString
 import qualified Data.Time as Time
 import qualified Data.UUID as UUID
 
-import           Data.Int (Int64)
+import           Data.Int (Int32, Int64)
 
 data PGBool
 data PGDate
@@ -71,7 +75,15 @@ pgStrictText = IPT.literalColumn . HPQ.StringLit . SText.unpack
 pgLazyText :: LText.Text -> Column PGText
 pgLazyText = IPT.literalColumn . HPQ.StringLit . LText.unpack
 
-pgInt4 :: Int -> Column PGInt4
+#if WORD_SIZE_IN_BITS == 32
+pgInt :: Int -> Column PGInt4
+pgInt = IPT.literalColumn . HPQ.IntegerLit . fromIntegral
+#elif WORD_SIZE_IN_BITS == 64
+pgInt :: Int -> Column PGInt8
+pgInt = IPT.literalColumn . HPQ.IntegerLit . fromIntegral
+#endif
+
+pgInt4 :: Int32 -> Column PGInt4
 pgInt4 = IPT.literalColumn . HPQ.IntegerLit . fromIntegral
 
 pgInt8 :: Int64 -> Column PGInt8